structopt, cleanup: compiler warnings

+++ src/transform/structOpt.cpp
@@ -108,7 +108,7 @@ StructOpt::transformLayout(llvm::AllocaInst & allocaInst, ValueToValueMapTy & tr
       auto vecBasePtr = transformMap[gep->getOperand(0)];
 
       std::vector<Value*> indexVec;
-      for (int i = 1; i < gep->getNumOperands(); ++i) {
+      for (size_t i = 1; i < gep->getNumOperands(); ++i) {
         indexVec.push_back(gep->getOperand(i));
       }
 
@@ -147,7 +147,7 @@ StructOpt::transformLayout(llvm::AllocaInst & allocaInst, ValueToValueMapTy & tr
 // repair phi nodes
   for (auto * phi : postProcessPhis) {
     auto * vecPhi = cast<PHINode>(transformMap[phi]);
-    for (int i = 0; i < phi->getNumIncomingValues(); ++i) {
+    for (size_t i = 0; i < phi->getNumIncomingValues(); ++i) {
       vecPhi->addIncoming(transformMap[phi->getIncomingValue(i)], phi->getIncomingBlock(i));
     }
   }
@@ -231,7 +231,7 @@ StructOpt::allUniformGeps(llvm::AllocaInst & allocaInst) {
     auto * gep = dyn_cast<GetElementPtrInst>(inst);
     if (gep) {
       // check whether all gep operands are uniform (except the baseptr)
-      for (int i = 1; i < gep->getNumOperands(); ++i) {
+      for (size_t i = 1; i < gep->getNumOperands(); ++i) {
         if (!getVectorShape(*gep->getOperand(i)).isUniform()) {
           IF_DEBUG_SO { errs() << "skip: non uniform gep: " << *gep << " at index " << i << " : " << *gep->getOperand(i) << "\n"; }
           return false;
@@ -245,7 +245,7 @@ StructOpt::allUniformGeps(llvm::AllocaInst & allocaInst) {
     auto * phi = dyn_cast<PHINode>(allocaUser);
     if (!phi) continue;
 
-    for (int i = 0; i < phi->getNumIncomingValues(); ++i) {
+    for (size_t i = 0; i < phi->getNumIncomingValues(); ++i) {
       auto * inVal = phi->getIncomingValue(i);
       if (isa<Instruction>(inVal) && !seen.count(inVal)) {
         IF_DEBUG_SO { errs() << "skip: alloca mixes with non-derived value " << *inVal << " at phi " << *phi << "\n"; }
@@ -366,7 +366,7 @@ StructOpt::vectorizeType(llvm::Type & scalarAllocaTy) {
 // finite aggrgate -> aggrgate of vectorized elemnts
   if (scalarAllocaTy.isStructTy()) {
     std::vector<Type*> elemTyVec;
-    for (int i = 0; i < scalarAllocaTy.getStructNumElements(); ++i) {
+    for (size_t i = 0; i < scalarAllocaTy.getStructNumElements(); ++i) {
       auto * vecElem = vectorizeType(*scalarAllocaTy.getStructElementType(i));
       if (!vecElem) return nullptr;
       elemTyVec.push_back(vecElem);
